# util/deephash: always keep values addressable