@dsnet 2022-08-10
There's a decent ~8% performance boost on the `Hash` benchmark which appears to be a more realistic benchmark of what we may hash. This occurs likely due to being able to use the fast-path more often.